refactor(control-plane): split sqlite migration tests
