+++ packages/cnes_infra/src/cnes_infra/control_plane/sqlite_schema.py
@@ -7,8 +7,7 @@ from typing import TYPE_CHECKING, Any
 
 from pydantic import BaseModel
 
-from cnes_domain.control_plane.entities import AccessRequest
-from cnes_domain.control_plane.enums import AccessRequestState, JobState
+from cnes_domain.control_plane.enums import JobState
 from cnes_domain.control_plane.errors import Conflict
 
 if TYPE_CHECKING:
@@ -458,41 +457,9 @@ def is_network_filesystem(path: Path) -> bool:
         if len(fields) >= 3 and resolved.is_relative_to(Path(fields[1])):
             matches.append((len(fields[1]), fields[2]))
     return bool(matches and max(matches)[1] in _NETWORK_FILESYSTEMS)
-def _migrate_schema(db: sqlite3.Connection) -> None:
-    job_columns = {row[1] for row in db.execute("PRAGMA table_info(job_creation_writes)")}
-    if "job_data" not in job_columns:
-        db.execute("ALTER TABLE job_creation_writes ADD COLUMN job_data TEXT")
-    db.execute(
-        "UPDATE job_creation_writes SET job_data = (SELECT data FROM jobs WHERE "
-        "jobs.tenant_id = job_creation_writes.tenant_id "
-        "AND jobs.job_id = job_creation_writes.job_id) WHERE job_data IS NULL")
-    run_columns = {row[1] for row in db.execute("PRAGMA table_info(runs)")}
-    if "unit_registry_data" not in run_columns:
-        db.execute("ALTER TABLE runs ADD COLUMN unit_registry_data TEXT")
-    access_columns = {row[1] for row in db.execute("PRAGMA table_info(access_requests)")}
-    if "creation_request_data" not in access_columns:
-        db.execute("ALTER TABLE access_requests ADD COLUMN creation_request_data TEXT")
-    for tenant_id, request_id, data in db.execute(
-        "SELECT tenant_id, request_id, data FROM access_requests "
-        "WHERE creation_request_data IS NULL"):
-        current = deserialize_model(data, AccessRequest)
-        original = current.model_copy(update={"state": AccessRequestState.PENDING,
-            "decided_by": None, "decided_at": None})
-        db.execute(
-            "UPDATE access_requests SET creation_request_data = ? WHERE "
-            "tenant_id = ? AND request_id = ?",
-            (serialize_model(original), tenant_id, request_id))
-    rows = db.execute(
-        "SELECT tenant_id, run_id FROM runs WHERE unit_registry_data IS NULL").fetchall()
-    for tenant_id, run_id in rows:
-        units = db.execute(
-            "SELECT data FROM run_units WHERE tenant_id = ? AND run_id = ? ORDER BY unit_id",
-            (tenant_id, run_id)).fetchall()
-        if units:
-            db.execute(
-                "UPDATE runs SET unit_registry_data = ? WHERE tenant_id = ? AND run_id = ?",
-                ("\x1e".join(row[0] for row in units), tenant_id, run_id))
 def initialize_schema(connect: Callable[[], sqlite3.Connection], path: Path) -> None:
+    from cnes_infra.control_plane.sqlite_migration import migrate_schema
+
     path.parent.mkdir(parents=True, exist_ok=True)
     connection = connect()
     try:
@@ -500,7 +467,7 @@ def initialize_schema(connect: Callable[[], sqlite3.Connection], path: Path) -> 
         if result is None or str(result[0]).lower() != "wal":
             raise _SQLiteWALUnavailable("sqlite_wal_unavailable")
         connection.executescript(_SCHEMA)
-        _migrate_schema(connection)
+        migrate_schema(connection)
         connection.commit()
     finally:
         connection.close()
